+++ packages/@pufflig/ps-nodes-config/src/core/forin.ts
@@ -23,7 +23,6 @@ export const forinNodeConfig: NodeConfig = {
       },
     ],
   },
-  parameters: [],
   outputs: [
     {
       id: "item",