+++ packages/@pufflig/ps-nodes-config/src/modifiers/handlebar_template/handlebar_template_chat.ts
@@ -4,7 +4,6 @@ export const handlebarTemplateChat: NodeConfig = {
   name: "Handlebar Template (Chat)",
   description: "",
   tags: ["modifier", "chat"],
-  parameters: [],
   outputs: [
     {
       id: "chat",