+++ packages/@pufflig/ps-nodes-config/src/modifiers/handlebar_template/handlebar_template_completion.ts
@@ -4,7 +4,6 @@ export const handlebarTemplateCompletion: NodeConfig = {
   name: "Handlebar Template (Completion)",
   description: "",
   tags: ["modifier", "text"],
-  parameters: [],
   outputs: [
     {
       id: "text",